Text size sanity checking for TEXT_ITEMS_GRID.
Oh dear, there was a bunch going on here.  Firstly the move from int
to long long int for ValueFromString() means that we were no longer
catching overflows (as we were C-style casting it back to int in many
places).  But even when the overflow is caught, it would run in to
wxWidgets' empty string bug while trying to log it.

Fixes https://gitlab.com/kicad/code/kicad/issues/12577

+++ common/dialogs/panel_setup_netclasses.cpp
@@ -336,7 +336,7 @@ bool PANEL_SETUP_NETCLASSES::TransferDataFromWindow()
     int row = 0;
 
     auto getCell =
-            [this]( int aRow, int aCol ) -> long long int
+            [this]( int aRow, int aCol ) -> int
             {
                 return m_frame->ValueFromString( m_netclassGrid->GetCellValue( aRow, aCol ) );
             };
